+++ src/web/Application.php
@@ -321,7 +321,7 @@ class Application extends \yii\web\Application
         // Override where Yii should find its asset deps
         $libPath = Craft::getAlias('@lib');
         $assetsPath = Craft::getAlias('@craft') . '/web/assets';
-        Craft::setAlias('@bower/jquery/dist', $libPath . '/jquery');
+        Craft::setAlias('@bower/jquery/dist', $assetsPath . '/jquery/dist');
         Craft::setAlias('@bower/inputmask/dist', $assetsPath . '/inputmask/dist');
         Craft::setAlias('@bower/punycode', $libPath . '/punycode');
         Craft::setAlias('@bower/yii2-pjax', $libPath . '/yii2-pjax');