+++ src/web/assets/jquery/webpack.config.js
@@ -1,0 +1,22 @@
+/* jshint esversion: 6 */
+/* globals module, require, __dirname */
+const CraftWebpackConfig = require('../../../../packages/craftcms-webpack/CraftWebpackConfig');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+const NODE_MODULES = __dirname + '/../../../../node_modules/';
+
+module.exports = new CraftWebpackConfig({
+    type: 'lib',
+    config: {
+        entry: {'entry': './entry.js'},
+        plugins: [
+            new CopyWebpackPlugin({
+                patterns: [
+                    {
+                        context: NODE_MODULES + '/jquery/dist',
+                        from: './jquery.*',
+                    }
+                ]
+            }),
+        ]
+    }
+});