+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added `craft\web\assets\inputmask\InputMaskAsset`.
+- Added `craft\web\assets\jquery\JqueryAsset`.
 
 ## 3.7.20 - 2021-11-06